Publish "worldping_check_info" metrics
This is used to reduce the number of labels applied to the metrics. A
single metric, worldping_check_info, has enough labels to perform a join
for the cases where it's necessary to filter by specific labels.

Also make sure that events get all the labels, because Loki does not
support joins.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>