Mark metrics stale before the scraper stops
If the scraper stops publishing data, Prometheus (Cortex) will "extend"
the last value for some amount of time (5 minutes), causing duplicate
metrics when the check is updated (particularly when changing things
that affect labels). To prevent this, push a stale marker for the most
recently collected metric, so that Prometheus knows that the metric
won't get any further updates.

This could be refined to look at the labels when a check is updated and
figure out if the metrics are affected or not.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>