Pass the authentication token via GPRC connection headers
Move the probe authetication from an argument in RPC calls to a
connection header at the client level. Since this is part of the
connection, all the services will get authenticated in this way.

If, in this future, it's necessary to have different authentication
tokens for different services, that can be implemented in the
GetRequestMetadata method.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>