Fix sidecar crashing when check is deleted
When adding instrumentation, I missed the fact that deleted checks are
mostly empty, so the "type" of the incoming check is invalid. We could
fix this on the worldping-api side, by retrieving the existing check and
sending that instead of the shell that the app sends to worldping-api.

Ask the scraper what kind of test it was running, and use that for the
metrics.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>