Add suport TCP checks
Add support for TCP checks. This is basically a check that creates a TCP
connection to a server, sends arbitrary strings and is able to verify
the responses. Optionally, it can use TLS for the connection.

Rework some of the code for scraper creation so that each type of check
is handled in a separate function.

Add support for TLSConfig both for tcp and http checks. It's possible to
verify the server's certificate and also present client certificates for
authentication. The crypto material is kept in temporary files, created
in the temporary directory ($TMPDIR if set, /tmp otherwise).

Also add BasicAuth to the HttpSettings, which was missing before.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>