Implement handling of CacheBustingQueryParamName
If CacheBustingQueryParamName is not empty, it is used as the query
parameter name passed to blackbox-exporter. The value of this parameter
is the current timestamp salted with the probe's name, hashed and
represented as hex.

This is optional, the name of the parameter must be specified by the
user to activate this functionality.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>