Handle disconnects
When the probe disconnects, don't die. Instead let GPRC handle the
reconnect, and once we get there, deal with the disparity between the
checks we have and the checks worldping-api thinks we have. Since an
arbitrary amount of time might have passed, deal with the fact that some
checks could have been updated, and some checks could have been deleted.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>