Fetch TenantInfo again if authentication fails
If we receive a 401 (Unauthorized) back from Prometheus or Loki, fetch
tenant information again as what we have in the cache might be stale.

If the information is stale, this is going to happen only once, since
the second time around it will succeed. If the information is stale, we
are going to get the same thing again, and we can drop the payload we
have because something *else* is wrong and we have no way of fixing it.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>