Feature: Implement test to check metric changes
Reorganize some of the code in the scraper, so that it's possible to run
a prober and collect metrics in a test.

Implement a test that runs each prober and compares the resulting
metrics against golden files, to verify that the metrics are not
changing underneath (for example if BBE gets upgraded).

Rename derived metrics so that they don't share the name with the
original counterparts. This requires the dashboards shipped with the
plugin to be updated!

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>