Fix: validate check and probe labels
We are not validating the labels that we are going to apply to the
metrics, and we eventually hit a call to MustRegister that causes the
probes to crash with bad metrics.

Add validation code for labels, that gets called from checks and probes.
This is going to cause the check to be rejected if it makes it to the
probe, but the intention is that it's rejected in the API or the
frontend, where it can be handled more gracefully. The API is going to
inherit this validation.

Also add TestCheckValidate to make sure we are calling these validations
for incoming checks.

Partially fixes issue #99 (because of required API and frontend fixes).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>