Communicate tenant info in changes
When then tenant remote information changes, the API will communicate
this to the probes in the changes stream.

The GetTenant service is still needed, because there might be changes
happening outside the API (for example, the user deleted a token), so
the probes still need to be able to query the tenant information on
their own.

Changing the name of the protocol buffer message is backwards
compatible.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>