Feature: provide access to accounting map
In order to make it possible for users of this code to validate that
new variations of check types have been added (or removed), provide
access to a copy of the accounting information via
GetAccountingClassInfo. Also provide access to the "accounting class" of
a check, which is the key used in that map. This makes it possible to
validate whether a check has switched classes if necessary.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>