Set user agent in outgoing HTTP requests
The user agent identifies the synthetic monitoring agent version, OS,
architecture, buildstamp and commit hash, and it points to the synthetic
monitoring product page on grafana.com so that people can more easily
identify requests as coming from Synthetic Monitoring.

The more explicit user agent version is also passed to Prometheus so
that it's easier to identify issues.

Closes #142 #117

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>