Remove quotes from ExecStart in .service file
Older versions of systemd handle quotes in ExecStart. At some point,
probably because people were doing it, systemd started handling quotes
differently. In particular, what's handling the word-splitting in the
arguments is systemd, not a shell, so it started looking at quotes to
prevent splits from ocurring within them. We don't have that case
because all the arguments should be single words without blanks in them.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>