Fix: remove direct dependency on github.com/grafana/loki
What we need from Loki is the logproto package, but depending on the
full Loki package is causing Go to pull a lot of problematic
dependencies that require hacks to keep them in check (see go.mod).

Simply copy the logproto package as-is (with the exception of a couple
of linter fixes) and use that instead.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>