Feature: disconnect agent from API on signal
If the agent receives a SIGUSR1 signal, it will disconnect from the API,
but it will continue running checks. This makes it possible for a new
version of the agent to connect to the API. When that happens, it's
possible to kill the running agent (SIGTERM or otherwise).

This should provide for a smoother upgrade process.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>