Feature: add log labels to log entries
This is a minimal change to support adding the log labels to the log
entries themselves.

In the future we would like to remove user-defined labels from the
labels and keep them in the log entries only in order to allow users to
add _more_ labels, but this change is not that.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>