Feature: add /ready endpoint for readiness probe
An HTTP request to the /ready endpoint will report whether the agent is
ready to run checks. Readyness is defined as having successfully
connected to the API once. After the initial connection, the agent will
start getting check updates from the API, and will continue running them
even if it's disconnected from the API.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>