Fix: correctly propagate check failure
The error metric is not getting updated because the scraper runner is
not reporting the failure as an error as expected. The failure is
correctly reported and recorded in user-visible metrics and logs, but
it's not accounted for in the agent's metrics, so it's possible to see
log entries for the checks entering the PASS state but no logs are
recorded for the FAIL state.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>