Feature: increase maximum number of user labels
Increase the maximum number of check labels to 10. Since Loki limits the
number of labels to 15, give priority to the internal labels (that are
used to provide some Synthetic Monitoring functionality), to probe
labels and then to check labels. All labels will be included as part of
the log entries themselves, so it's still possible to access them via
LogQL queries.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>