Feature: Implement alternative ICMP prober
Change the ICMP implementation to one that can send out multiple packets
per run and report min and max values for the roundtrip.

The underlying code ignores the DontFragment flag, but that should be OK
as we are not exposing settings to make a difference in that case.

This adds a check to see if privileged sockets are required. Privileged
is a property of the environment, so cache the result for the rest of
the life of the agent.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>