Fix: Re-enable request recursion
BBE v0.20.0 added a `recursive` option that defaults to true which sets
the recursion desired bit in the DNS request. Since we don't use the
configuration file but instead build the module's configuration directly
in code, we are not picking up the YAML defaults. When we switched from
v0.19.0 to v0.20.0, the Recursion field defaults to false, causing DNS
requests to be non-recursive, i.e. they MUST be resolved by the
specified target server.

I'm still scratching my head as to why Google's public DNS server
(dns.google, 8.8.8.8) is in fact resolving grafana.com when its NS
server is ns1 (nsone.net):

    $ dig +norecurse @8.8.8.8 grafana.com NS

    ; <<>> DiG 9.16.8-Ubuntu <<>> +norecurse @8.8.8.8 grafana.com NS
    ; (1 server found)
    ;; global options: +cmd
    ;; Got answer:
    ;; ->>HEADER<<- opcode: QUERY, status: NOERROR, id: 36686
    ;; flags: qr ra; QUERY: 1, ANSWER: 4, AUTHORITY: 0, ADDITIONAL: 1

    ;; OPT PSEUDOSECTION:
    ; EDNS: version: 0, flags:; udp: 512
    ;; QUESTION SECTION:
    ;grafana.com.                   IN      NS

    ;; ANSWER SECTION:
    grafana.com.            3047    IN      NS      dns1.p09.nsone.net.
    grafana.com.            3047    IN      NS      dns2.p09.nsone.net.
    grafana.com.            3047    IN      NS      dns3.p09.nsone.net.
    grafana.com.            3047    IN      NS      dns4.p09.nsone.net.

    ;; Query time: 64 msec
    ;; SERVER: 8.8.8.8#53(8.8.8.8)
    ;; WHEN: Mon Mar 28 10:17:01 CST 2022
    ;; MSG SIZE  rcvd: 129

    $ dig +norecurse @8.8.8.8 grafana.com A

    ; <<>> DiG 9.16.8-Ubuntu <<>> +norecurse @8.8.8.8 grafana.com A
    ; (1 server found)
    ;; global options: +cmd
    ;; Got answer:
    ;; ->>HEADER<<- opcode: QUERY, status: NOERROR, id: 18200
    ;; flags: qr ra; QUERY: 1, ANSWER: 1, AUTHORITY: 0, ADDITIONAL: 1

    ;; OPT PSEUDOSECTION:
    ; EDNS: version: 0, flags:; udp: 512
    ;; QUESTION SECTION:
    ;grafana.com.                   IN      A

    ;; ANSWER SECTION:
    grafana.com.            430     IN      A       34.120.177.193

    ;; Query time: 52 msec
    ;; SERVER: 8.8.8.8#53(8.8.8.8)
    ;; WHEN: Mon Mar 28 10:19:06 CST 2022
    ;; MSG SIZE  rcvd: 56

Cloudflare's public DNS servers behave in the same way for A records,
but not for NS records. In the later case they will return an answer
only if the recursion desired bit is set.

On the other hand, the same NS and A queries fail for raintank.org when
using Google's public DNS servers.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>