Feature: add a connection health ping
Try to send a ping to the API. If it succeeds, keep doing it on a
regular interval, so that the API can determine whether the client is
still alive or not. If the API reports that it doesn't support the
method, give up and assume that we are talking to an older API.

On the API side, it should receive the ping and disconnect the client if
it fails to communicate on a regular interval. By sending back a reply
we make sure that the client can also make sure the API is alive.

The 90 second interval is chosen so that it's possible to set up a
slightly longer timeout elsewhere just in case.

This is necessary because when going thru a proxy, we have observed that
in some cases the proxy keeps the connection to the API up even if the
client itself has dropped the connection.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>