Fix: enable HTTP2 by default
After upgrading to github.com/prometheus/common v0.34.0, HTTP2 is
enabled by default, but the problem is that the way this is done
requires setting a value to true. Since this is a new field, in Go it
defaults to false, so the "default" has to be implemented by hand.

We don't expose a setting to disable this. Yet?

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>