Fix: correctly propagate conectivity errors
The current version of the code is canceling the context when there's a
ping error, so that the function waiting for changes stops. This causes
the rest of the logic to observe a "context canceled" error, which it
interprets as the program shutting down and never trying to reconnect.

Propagate context.Canceled up the call stack, and handle it uniformly in
a single location. Group transient errors under its own type, so that
the handling logic is uniform, too. Use an errror group to collect the
error conditions from the goroutines (the first error condition is
reported back up the call stack, everything else is discarded).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>