build: prevent invalid os/arch combinations
This adjusts the Makefile to prevent crossbuilding for invalid OS/arch
targets. After this, `make build` will create binaries for `linux/arm`,
`linux/arm64`, `linux/amd64` and the local platform, if different.

Originally, it would build for all combinations of:
    GOOS={current, linux}
    GOARCH={current, amd64, arm, arm64}

This could cause the build to fail when run on an OS other than Linux.
For example, when run under macOs, it will result in an attempt to build
a `darwin/arm` binary, which is not a platform supported by Go.