Fix: handle connection state changes
While trying to figure out why we are seeing a GOAWAY HTTP2 frame coming
from the "API" (actually nginx), Adrian found out that the only way to
handle that is by watching the connection state changes.

Add another goroutine that is in charge of monitoring the state changes.
If the connection goes to IDLE or SHUTDOWN, report an error from the
goroutine so that it breaks out of the loop. This causes the agent to
reconnect.

Our best understanding of the issue is that nginx has a hard limit on
the number of requests that can be made over a single connection and
when it reaches that limit, it sends a GOAWAY frame. The pings that we
are sending to keep the connection alive are making this worse becase we
send one every 90 seconds and the current limit in nginx is 100 (that
gives us 9000 = 2.5 hours, not counting the additional requests that are
part of the regular communication between the API and the agent, so
that's the upper limit; that seems to match observed behavior).

See https://github.com/grafana/deployment_tools/pull/47478

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>