Fix: WANTED_OSES / WANTED_ARCHES was removed, use PLATFORMS
A while back the use of the WANTED_OSES and WANTED_ARCHES variables was
removed in favor of buidling a single list using PLATFORMS, which has a
default value.

When this was fixed, the scripts/package/goreleaser-build wrapper was
broken, and we didn't notice because we didn't do a release. Later we
didn't notice because we ourselves use the docker images, which are
unaffected by this.

Fix this by passing PLATFORMS in the wrapper.

Fixes: #367
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>