Feature: update BBE to version 0.23.0
This exposes a new metric, probe_dns_query_succeeded, which is set to 1
after the query is sent to the DNS server and the server responds. This
allows the user to differentiate between cases when the server failed to
respond (and the probe is reported as having failed) vs the server
responding with somehting that causes the probe to fail. In other words,
probe_success set to 0 but probe_dns_query_succeeded not set means the
DNS server failed to respond, while probe_success set to 0 and
probe_dns_query_succeeded set to 1 means the server responded but the
probe failed for a different reason (e.g. the record does not exist).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>