Fix: truncate long label values
We generate some label values, so they are not being subject to the
constraints the labels provided by the user are.

In order to prevent the metrics from being dropped by Prometheus because
the label values exceed the limit, truncate the values that exceed the
limit, and replace the last three bytes by '...' to indicate that
truncation took place.

For the specific case of certificates with a really long list of Subject
Alternate Names (SAN), they end up in a label. It's a list that looks
like `a.example.org,b.example.org,c.example.org`. We _could_ split that
list and push each individual value as a separate label value just to
avoid hitting the limit. The problem with that approach is that we have
seen certificates with over 700+ SANs in the wild, so this strategy would
push out 700+ metrics.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>