Fix: don't use 0 in subject and condition enums
Keep 0 as the "default" value. For some assertion types that means not
used, and for others it means a different value will be used
(RESPONSE_BODY for subject, CONTAINS for condition).

This is backwards-compatible: if a check is sent with the value 0, it
uses the corresponding value. Future checks should use the non-zero
values. At some point in the future, we can update the code to error out
if a 0 is used.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>