Feature: use expression to match specific headers in multiHTTP
In order to be able to match specific headers in multiHTTP checks, we
use the existing `expression` field to specify the header name to match.
Other than this, everything works as before. There's a fallback in
place, so that in case we have a check _without_ an expression, it uses
the old behavior.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>