Use local IDs in scraper, global IDs in pusher
In order to avoid confusion as to which kind of ID is expected where,
introduce a model package that takes a synthetic_monitoring.Check and
produces a wrapped one that has the IDs decoded.

Introduce a model.GlobaID type to make it explicit which kind of ID is
expected. In the scraper use local IDs, except for the lists of checks,
which need to use a global one. In the publisher, use global IDs.

This allows the compiler to flag misuses.