Fix: flip order of assertions and variable definitions
While it probably makes sense to assert before trying to capture values
from the response, defining variables before assertions allows the user
to use those variables in the assertions without making a second record.
A possible use case is capturing a value from the response and verifying
that that value is present (or not present) elsewhere in the response.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>