Add jitter to tenant's maxLifetime
In order to avoid all tenants from trying to refresh at about the same
time, add some jitter to the maxLifetime they use. This should spread
the updates as time passes.

This has the risk of having some tenants update more often than others,
but that should not happen often. It still has the problem that during
initialization all tenants are going to be updated at about the same
time.

I think the longer term solution is to not to tear down the publisher on
a regular basis, but simply update the client its using.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>