Add methods to convert to and from strings for MultiHTTP enums
In order to provide a more consistent and better experience to users,
add methods to convert back and fort between enum values and their
string representations.

The code is generated using enumer. A lot of the noise comes from adding
support for that (and it's still incomplete, it will need revisiting).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>