Fix: add a `name` label to metrics
By adding a `name` tag to each request, we get that in the result as a
`name` label with the corresponding value.

Since the [k6
documentation](https://k6.io/docs/using-k6/http-requests/#url-grouping)
says that setting `name` should not affect `url` but [in reality it
does](https://github.com/grafana/k6/issues/3435), implement a nasty
workaround to preserve the URL. If there's a tag with the special name
`__raw_url__`, use that as the actual URL that was requested.

Modify the generated script so that we add both the `name` tag (set to a
sequential number based on the position of the request in the entries
list) and a `__raw_url__` tag (workaround). Then in the output
extension, include `name` in the metric labels and convert `__raw_url__`
to `url`.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>