Fix: make the k6 runner timeout configurable (#554)
When a prober is created it gets a context that has the check's timeout
added as the deadline for the context. The timeout is measured between
the point where the probe is started and the point where it finishes
running. For the MultiHTTP case, we _add_ the k6 start up time to that,
because this behavior is consistent with the way the other checks work.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>
Co-authored-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>