Export max agent labels query functions
Defines public functions that return the maximum number of labels that
can be set for a metric and a log stream. These are static per agent
version, therefore are defined as private constants, but are exposed
through a public function to make it more "intuitive" to understand that
they may vary between agent versions, and also to allow encapsulating
logic in these methods in the future if required.