Fix: make label names uniform
Our use of labels names in logs and metrics is not uniform and this is
causing a lot of pain.

We have `foo_id`, `fooID` and other variations.

Pick one.

It's `fooId` from now on.

It's going to be painful to change everywhere else to match this, but
it's better than dealing with the existing pain.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>