Add Telemetry pkg
Adds the telemetry package responsible for sending telemetry, currently
for executions and executios duration, to the API.
For each region that has tenants executing checks in the agent
instance there will be an associated RegionPusher, which holds the
telemetry data for that specific region, propagated through the
Telemeter component. Each RegionPusher is executed in a loop which
pushes the data to the API at specified time intervals. Additionally it
will also push the currently accumulated data when the global context is
done, before the agent stops its execution.