Expose agent label constants (#585)
* Refactor TestValidateMetrics

Extracts the probe setup functions so test is simplified and these can
be reused for other scraper tests.

* Export max agent labels query functions

Defines public functions that return the maximum number of labels that
can be set for a metric and a log stream. These are static per agent
version, therefore are defined as private constants, but are exposed
through a public function to make it more "intuitive" to understand that
they may vary between agent versions, and also to allow encapsulating
logic in these methods in the future if required.

* Add TestValidateLabels

The purpose of this test is to verify that the currently defined
constants for max metric and log labels set by the agent still apply and
have not changed due to modifications, in which case the constants
should be updated.

* Remove max check labels verification

This should now be done in the API taking into account the particular
tenant limits.