Feature: promote adhoc to permanent feature
This has caused a number of questions recently because the "test"
functionality requires the "adhoc" feature to be enabled.

We have had this feature for a long time now and we are actively using
it. Make it a permanent feature by removing the checks for whether or
not it's enabled. Keep the feature flag in case someone is passing it
on the command line to avoid breaking those set ups.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>