Prevent race condition on telemeter's AddExecution
If the region has no pusher initialized, acquiere the W lock just after
the R lock has been released, so we avoid possible overlapping work
between two calls that were trying to add an execution for the same
region when there was no pusher for that region.
Once the W lock is acquiered, verify that no concurrent call acquiered
the lock just before and already initialized a pusher for the region.
This could make the "second call" to replace the just initialized pusher
for the region, and lose execution data.