Switch to using buf to manage protobuf code
This required some rework around code generation (e.g. we use enumer to
generate strings out of enums, enumer is part of grafana-build-tools, so
that required some adjusting to work correctly).

Because of the above, this is adding a `scripts/docker-run` script that
is used to run tools from grafana-build-tools consistenly.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>