Use github.com/grafana/loki/pkg/pusher instead of logproto
For a long time, we have used an internal copy of Loki's logproto
package. At some point Loki moved that to a "push" package, and heavily
modified "logproto". That push package is a separate Go module, so we
don't need to worry about extra dependencies.

Use that "push" package instead.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>