Experimental: increase max frequency to 1 hour (#645)
Allow the maximum interval between successive check runs to be up to 1
hour. This will break some dashboards, because there are built-in
assumptions about the maximum value being 2 minutes. In particular, it's
going to be necessary to adjust the value we use for ranges, which is
hard-coded to 5 minutes. If we change the queries, we still need to
adjust values in Grafana, to prevent sampling values at intervals less
than the frequency. This is going to be particularly visible in queries
that use `increase`, `rate` and similar functions. Because of that, this
is mostly about adding backend support for [issue #318](https://github.com/grafana/synthetic-monitoring-app/issues/318).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>