Fix: report duration from script
Right now we are reporting the time it takes to run a script, including
the time it takes for k6 to actually start running the script, plus all
other kinds of overhead.

Modify the sm extension to report the script duration, and report that
number as probe_duration_seconds. We are still reporting
probe_script_duration_seconds as well. This will get fixed in a future
change.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>