Output unadorned version tag for releases (#862)
There's stuff out there with pretty weird ideas as to what constitutes a
"correct" version number. We have been trying to solve this problem of
"incorrect versioning" for over 30 years with no success. You have to
wonder why that is.

For the specific case of releases, in addition to using vX.Y.Z-0-sha,
also output a vX.Y.Z tag. That allows stuff to be happy about seeing
something that matches its expectations in terms of version numbers. We
still need to output something else in the n != 0 case, because we need
to be able to differentiate between intermediate builds. It's incorrect
in semver terms, because those are prereleases that sort _before_
vX.Y.Z, not after it. To do it correctly, we would have to transform
vX.Y.Z to vX.Y.(Z+1) and append the prerelease bit to that. Since that
would become confusing really fast (in the face of existing practice),
let's not do that. As long as nothing tries to "upgrade" vX.Y.Z-n-sha to
vX.Y.Z, we shoud be OK.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>