Feature: add retries to ICMP prober
Add a retry mechanism to the ICMP prober. If the packet count is not
set, use 3 packets instead of 1, and require that at least 1 of them
gets a response. This should cover most reasonable scenarios where
retries are needed.

Note that this will send out 3 packets regardless of whether 1 is
already received, so this has an impact on the probe_duration_seconds
metric (because the probe takes longer to run).

If the time for an individual packet is important, use
`sum(probe_icmp_duration_seconds)` instead. That metric is split by
phases, and the `rtt` phase is the average round-trip-time for all the
received packets. The other two phases, `setup` and `resolve` only run
once.

The resolve part of this also retries, in the event that the DNS server
fails to produce a response and there's an indication that the request
should be retried (e.g. if the response is "not found", that is not
retried).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>