Fix: allow probers to provide a duration value
Some probers need to override the duration that is reported, mostly
because we use probe_duration_seconds as the "latency" of the probe,
which isn't exactly correct. `probe_duration_seconds` was meant to be
used as the duration of the probe: how long did it take for this probe
to run. Capture that value in the logs for debugging purposes, and use
`probe_duration_seconds` as if it meant the time it takes the target to
produce a response, including *some* setup, but not e.g. the overhead
introduced by retries or the time it takes for external programs to spin
up.

This is OK to change since these are all internal packages.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>