Add script metadata
In order for the runner to know the type of the script that it's dealing
with, add a "metadata" field to the JSON representation of the request,
which contains a "type" field specifying that information.

Also pass check_id and tenant_id, which will allow us to identify checks
causing issues.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>